backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
pydantic==2.10.6
paho-mqtt==2.1.0
cbor2==5.6.5
python-multipart==0.0.20
pytest==8.3.4
httpx==0.28.1
